chat_system/requirements.txt
Django==1.8
djangorestframework==3.0.4
channels==0.12.0
django-channels-presence==0.0.7
django-environ==0.3.0
MySQL-python==1.2.5
django-braces==1.8.0
celery==4.0.2
django-celery-beat==1.0.1
asgi-redis==1.4.0
asgiref==1.1.2
redis==2.10.5
requests==2.17.3
bpython==0.16
